requirements.txt
tensorflow
pandas
numpy
alpaca-trade-api
matplotlib
ipykernel
scikit-learn
alpaca-py
python-dateutil